backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.121.3
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
pydantic==2.12.4
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-multipart==0.0.20

# AI & ML
google-generativeai==0.8.5
faiss-cpu==1.9.0.post1
numpy==2.2.1

# Database
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
email-validator==2.1.0

# Testing
pytest==9.0.1
pytest-asyncio==1.3.0
httpx==0.28.1

